contrib.task tests: Nailed coverage with more tests.
Seems we're using mock to override some methods so 'coverage' was quite
right in telling me that these test cases weren't being considered.

Solution: add some more test cases.  It seems this has uncovered a bug
in _exception, when there's no logger it tries (the wrong way) to raise
an exception with a traceback.

Now to fix that problem.