signalslot.contrib.task: Use six.reraise
raise syntax changed between Python 2.x and 3.x, and to my annoyance,
there is no compatible syntax that *BOTH* agree with.  Even if wrapped
in an IF statement, or put in separate modules, the tentacles of Python
3.x still manage to find that Python 2 code and tries to parse it.

Since we already drag in 'six', there's a solution, use *THAT* and its
reraise function to re-raise the exception.  The test now passes on both
Python 2.7 and 3.4.